fix: add lowerCase headers
