+++ task-1/csv/output.txt
@@ -1,0 +1,5 @@
+{"book":"The Compound Effect","author":"Darren Hardy","price":"9,48"}
+{"book":"The 7 Habits of Highly Effective People","author":"Stephen R. Covey","price":"23,48"}
+{"book":"The Miracle Morning","author":"Hal Elrod","price":"21,34"}
+{"book":"Influence: The Psychology of Persuasion","author":"Robert B. Cialdini","price":"12,99"}
+{"book":"The ONE Thing","author":"Gary Keller","price":"11,18"}
